Possible upgrade for 0.7.9
