+++ src/templates/extra_senses.html
@@ -1,19 +1,19 @@
 <hr/>
-<p class="notes">{{localize "CONVIS.ExtraSenseInstructions"}}</p>
+<p class="notes">{{localize "conditional-visibility.ExtraSenseInstructions"}}</p>
 <div class="form-group">
-    <label>{{localize "CONVIS.blindsight"}}</label>
+    <label>{{localize "conditional-visibility.blindsight"}}</label>
     <input type="checkbox" {{ checked blindsight}} name="flags.conditional-visibility.blindsight"/>
-    <label>{{localize "CONVIS.truesight"}}</label>
+    <label>{{localize "conditional-visibility.truesight"}}</label>
     <input type="checkbox" {{ checked truesight}} name="flags.conditional-visibility.truesight"/>
 </div>
 <div class="form-group">
-    <label>{{localize "CONVIS.devilssight"}}</label>
+    <label>{{localize "conditional-visibility.devilssight"}}</label>
     <input type="checkbox" {{ checked devilssight}} name="flags.conditional-visibility.devilssight"/>
-    <label>{{localize "CONVIS.tremorsense"}}</label>
+    <label>{{localize "conditional-visibility.tremorsense"}}</label>
     <input type="checkbox" {{ checked tremorsense}} name="flags.conditional-visibility.tremorsense"/>
 </div>
 <div class="form-group">
-    <label>{{localize "CONVIS.seeinvisible"}}</label>
+    <label>{{localize "conditional-visibility.seeinvisible"}}</label>
     <input type="checkbox" {{ checked seeinvisible}} name="flags.conditional-visibility.seeinvisible"/>
     <label>&nbsp;</label>
     <input type="checkbox" style="visibility: hidden;" disabled name="conditionalVisibility_justtolineup"/>
