+++ src/templates/help_dialog.html
@@ -1,12 +1,12 @@
 <h2>Conditional Visibility</h2>
-<p class="notes">{{localize "CONVIS.help.gamesystem"}} : {{gamesystem}}, {{localize "CONVIS.help.hasStealth"}}: {{hasStealth}}</p>
+<p class="notes">{{localize "conditional-visibility.help.gamesystem"}} : {{gamesystem}}, {{localize "conditional-visibility.help.hasStealth"}}: {{hasStealth}}</p>
 {{#if hasStealth}}
-<p class="notes">{{localize "CONVIS.settings.autoStealth.name"}} : {{autoStealth}}</p>
+<p class="notes">{{localize "conditional-visibility.settings.autoStealth.name"}} : {{autoStealth}}</p>
 {{/if}}
 <table>
     <thead>
         <tr>
-            <th colspan="2">{{localize "CONVIS.help.availableconditions"}}</th>
+            <th colspan="2">{{localize "conditional-visibility.help.availableconditions"}}</th>
         </tr>
     </thead>
     <tbody>
@@ -18,4 +18,4 @@
         {{/each}}
     </tbody>
 </table>
-<p class="notes">{{localize "CONVIS.help.readmeat"}} <a href="https://github.com/gludington/conditional-visibility/blob/master/README.md">README</a></p>    
+<p class="notes">{{localize "conditional-visibility.help.readmeat"}} <a href="https://github.com/gludington/conditional-visibility/blob/master/README.md">README</a></p>
