+++ src/templates/stealth_hud.html
@@ -1,5 +1,5 @@
 <div class="form-group">
-    <label>{{localize "CONVIS.stealthroll"}}</label>
-    <input id="CONVIS.stealth" value="{{initialValue}}"/>
+    <label>{{localize "conditional-visibility.stealthroll"}}</label>
+    <input id="conditional-visibility.stealth" value="{{initialValue}}"/>
 </div>
-<p class="notes">{{localize "CONVIS.stealthrolldescription"}}</p>
+<p class="notes">{{localize "conditional-visibility.stealthrolldescription"}}</p>
