+++ src/templates/version_popup.html
@@ -1,4 +1,4 @@
 <h2>Conditional Visibility {{version}}</h2>
-<p class="notes">{{localize "CONVIS.popup.version0910.1"}}</p>
-<p class="notes">{{localize "CONVIS.popup.version0910.2"}}</p>
-<p class="notes">{{localize "CONVIS.popup.version0910.3"}}</p>
+<p class="notes">{{localize "conditional-visibility.popup.version0910.1"}}</p>
+<p class="notes">{{localize "conditional-visibility.popup.version0910.2"}}</p>
+<p class="notes">{{localize "conditional-visibility.popup.version0910.3"}}</p>
